+++ SpellActivationOverlay.lua
@@ -14,6 +14,7 @@ local useTimer = true;
 local useSound = false;
 
 function SpellActivationOverlay_OnLoad(self)
+	SAO_Frame = self;
 	SAO.Frame = self;
 	SAO.ShowAllOverlays = SpellActivationOverlay_ShowAllOverlays;
 	SAO.HideOverlays = SpellActivationOverlay_HideOverlays;
@@ -174,14 +175,14 @@ function SpellActivationOverlay_OnEvent(self, event, ...)
 		-- end
 	end
 	if ( not self.disableDimOutOfCombat ) then
-		if ( event == "PLAYER_REGEN_DISABLED" ) then
+		if ( event == "PLAYER_REGEN_DISABLED" and self.inPseudoCombat ~= true ) then
 			self.combatAnimOut:Stop();	--In case we're in the process of animating this out.
 			self.combatAnimIn:Play();
 			for _, overlay in ipairs(self.combatOnlyOverlays) do
 				overlay.combat.animOut:Stop();
 				SpellActivationOverlayFrame_PlayCombatAnimIn(overlay.combat.animIn);
 			end
-		elseif ( event == "PLAYER_REGEN_ENABLED" ) then
+		elseif ( event == "PLAYER_REGEN_ENABLED" and self.inPseudoCombat ~= false ) then
 			self.combatAnimIn:Stop();	--In case we're in the process of animating this out.
 			self.combatAnimOut:Play();
 			for _, overlay in ipairs(self.combatOnlyOverlays) do
@@ -354,8 +355,19 @@ function SpellActivationOverlay_ShowOverlay(self, spellID, texturePath, position
 	if useSound and (autoPulse or forcePulsePlay) then
 		overlay.soundHandle = SAO:PlaySpellAlertSound();
 	end
+	-- Circumvent a bug with pulse animations
+	-- First, hide the overlay temporarily
+	-- Then, very quickly (but not too quickly) start the pulse animation sooner, which kind of fixes the animation issue
+	-- If we do not do this, we might see a weird flash for a brief moment at the very beginning of overlay:Show()
+	if ( combatOnly ) then
+		overlay.animIn.alpha1:SetToAlpha(0.01);
+		overlay.animIn.alpha2:SetFromAlpha(0.01);
+	else
+		overlay.animIn.alpha1:SetToAlpha(0.5);
+		overlay.animIn.alpha2:SetFromAlpha(0.5);
+	end
 	overlay:Show();
-	if ( forcePulsePlay ) then
+	if ( forcePulsePlay and not overlay.pulse:IsPlaying() ) then
 		overlay.pulse:Play();
 	end
 	overlay.pulse.autoPlay = autoPulse;
@@ -389,6 +401,13 @@ function SpellActivationOverlay_ShowOverlay(self, spellID, texturePath, position
 			overlay.combat.animOut:Stop();
 			SpellActivationOverlayFrame_PlayCombatAnimIn(overlay.combat.animIn);
 		end
+	end
+end
+
+function SpellActivationOverlay_DumpCombatOnlyOverlays()
+	SAO:Info(Module, "Listing combat-only overlays ("..#SAO.Frame.combatOnlyOverlays.." item"..(#SAO.Frame.combatOnlyOverlays == 1 and "" or "s")..")");
+	for i, overlay in pairs(SAO.Frame.combatOnlyOverlays) do
+		SAO:Info(Module, "combat-only-overlay["..i.."] location == "..overlay.position..", spell ID = "..overlay.spellID.." "..(GetSpellInfo(overlay.spellID) or ""));
 	end
 end
 
@@ -531,11 +550,13 @@ function SpellActivationOverlayTexture_TerminateOverlay(overlay)
 	overlay.soundHandle = nil;
 
 	-- Hide the overlay and make it available again in the pool for future use
+	overlay.mask:SetAlpha(0); -- Set alpha to zero, because next time we want to start unnoticed
 	overlay.mask:SetScale(1); -- Reset scale, in case a previous animation shrank it to 0.01
 	overlay.endTime = nil; -- Reset endTime, to avoid excessive optimizations when re-using this overlay
 	overlay:Hide();
 	tDeleteItem(overlayParent.overlaysInUse[overlay.spellID], overlay);
 	tinsert(overlayParent.unusedOverlays, overlay);
+	tDeleteItem(overlayParent.combatOnlyOverlays, overlay);
 end
 
 function SpellActivationOverlayFrame_OnTimeoutFinished(anim)
@@ -651,14 +672,15 @@ function SpellActivationOverlayTexture_OnFadeInFinished(animGroup)
 	SAO:Trace(Module, "SpellActivationOverlayTexture_OnFadeInFinished "..tostring(animGroup));
 	local overlay = animGroup:GetParent();
 	overlay:SetAlpha(1);
-	if ( overlay.pulse.autoPlay ) then
+	if ( overlay.pulse.autoPlay and not overlay.pulse:IsPlaying() ) then
 		overlay.pulse:Play();
 	end
 end
 
 function SpellActivationOverlayTexture_PreStartPulse(anim)
+	SAO:Trace(Module, "SpellActivationOverlayTexture_PreStartPulse "..tostring(anim));
 	local overlay = anim:GetParent():GetParent();
-	if ( overlay.pulse.autoPlay ) then
+	if ( overlay.combatOnly and overlay.pulse.autoPlay and not overlay.pulse:IsPlaying() ) then
 		overlay.pulse:Play();
 	end
 end
@@ -684,6 +706,22 @@ function SpellActivationOverlayFrame_OnFadeInFinished(anim)
 			end
 		end
 	end
+end
+
+function SpellActivationOverlayFrame_OnEnterCombat(anim)
+	SAO:Trace(Module, "SpellActivationOverlayFrame_OnEnterCombat "..tostring(anim));
+	-- Combat has started, or pseudo-started
+	-- A pseudo-start happens when a proc was just triggered and the effect is visible shortly
+	-- In this case, the player may or may not be in combat, but overlays will be visible as if in combat
+	local frame = anim:GetParent();
+	frame.inPseudoCombat = true;
+end
+
+function SpellActivationOverlayFrame_OnLeaveCombat(anim)
+	SAO:Trace(Module, "SpellActivationOverlayFrame_OnLeaveCombat "..tostring(anim));
+	-- Combat has finished, or pseudo-finished (see SpellActivationOverlayFrame_OnEnterCombat)
+	local frame = anim:GetParent();
+	frame.inPseudoCombat = false;
 end
 
 function SpellActivationOverlayFrame_SetForceAlpha1(enabled)